feat(ProductDetail and Detail): modularize comp. and add modal to delete function
