+++ PF_Client/src/views/ProductDetail/ProductDetail.jsx
@@ -1,13 +1,24 @@
 import React, { useEffect, useState } from "react";
 import style from "./ProductDetail.module.css";
 import { useSelector, useDispatch } from "react-redux";
+import { Link } from 'react-router-dom';
 import {
   deleteProduct,
   getProductById,
   getProductQuestions,
   setProductChange,
 } from "../../Redux/actions";
+import { useDisclosure,
+  Button,
+ Modal,
+ ModalOverlay,
+ ModalContent,
+ ModalHeader,
+ ModalFooter,
+ ModalBody,
+ ModalCloseButton,} from "@chakra-ui/react";
 import { NavBar } from "../../components/NavBar/NavBar";
+import Detail from "../../components/Detail/Detail";
 export const ProductDetail = () => {
   const details = useSelector((state) => state.details);
   const dispatch = useDispatch();
@@ -15,7 +26,7 @@ export const ProductDetail = () => {
   const questions = details.questions;
   const [currentImg, setCurrentImg] = useState(0);
   //Variable provisoria que a futuro deberá llegar desde el estado global
-  const userId = 3;
+  const userId = 1;
   const [edit, setEdit] = useState({
     e: false,
     s: "none",
@@ -63,196 +74,18 @@ export const ProductDetail = () => {
       {details.length !== 0 ? (
         <div>
           <h1> Detalle del producto </h1>
-          {userId === details.userid ? (
-            <div>
-              <button onClick={handleEdition}>Habilitar Edición 🖊</button>
-              <button onClick={handleDelete}>Eliminar</button>
-            </div>
-          ) : (
-            ""
-          )}
-          <table className={style.detailTable}>
-            <tr>
-              <td>
-                <button
-                  name="leftBtn"
-                  className={
-                    currentImg === 0
-                      ? style.ArrowButtonDisabled
-                      : style.ArrowButton
-                  }
-                  onClick={() =>
-                    currentImg === 0
-                      ? setCurrentImg(0)
-                      : setCurrentImg(currentImg - 1)
-                  }
-                >
-                  {"◀"}
-                </button>
-              </td>
-              <tr>
-                <td className={style.box} rowSpan="6">
-                  <img
-                    className={style.FlagImg}
-                    src={details.image[currentImg]}
-                    alt={`imagen del producto ${details.name}`}
-                  />
-                </td>
-                {/* <td>{edit.s === "image" ?
-              (
-                <span>
-                  <input
-                    type="text"
-                    name="image"
-                    onChange={InputHandler}
-                  />
-                  <button onClick={SendCange}>OK</button>
-                </span>
-              ) : ''}
-              </td>
-              <td>
-                    {edit.e ? (
-                      <button name="image" onClick={EditionPDetail}>
-                        {" "}
-                        🖊{" "}
-                      </button>
-                    ) : (
-                      ""
-                    )}
-                  </td> */}
-              </tr>
-
-              <td>
-                <button
-                  name="rightBtn"
-                  className={
-                    currentImg === details.image.length - 1
-                      ? style.ArrowButtonDisabled
-                      : style.ArrowButton
-                  }
-                  onClick={() =>
-                    currentImg === details.image.length - 1
-                      ? setCurrentImg(currentImg)
-                      : setCurrentImg(currentImg + 1)
-                  }
-                >
-                  {"▶"}
-                </button>
-              </td>
-              <td className={style.tdLeft}>
-                <tr>
-                  <td>
-                    Producto:
-                    {edit.s === "name" ? (
-                      <span>
-                        <input
-                          type="text"
-                          name="name"
-                          onChange={InputHandler}
-                        />
-                        <button onClick={SendCange}>OK</button>
-                      </span>
-                    ) : (
-                      details.name
-                    )}
-                  </td>
-                  <td>
-                    {edit.e ? (
-                      <button name="name" onClick={EditionPDetail}>
-                        {" "}
-                        🖊{" "}
-                      </button>
-                    ) : (
-                      ""
-                    )}
-                  </td>
-                </tr>
-                <tr>
-                  <td>
-                    Descripcion:
-                    {edit.s === "description" ? (
-                      <span>
-                        <input
-                          type="text"
-                          name="description"
-                          onChange={InputHandler}
-                        />
-                        <button onClick={SendCange}>OK</button>
-                      </span>
-                    ) : (
-                      details.description
-                    )}
-                  </td>
-                  <td>
-                    {edit.e ? (
-                      <button name="description" onClick={EditionPDetail}>
-                        {" "}
-                        🖊{" "}
-                      </button>
-                    ) : (
-                      ""
-                    )}
-                  </td>
-                </tr>
-                <tr>
-                  <td>
-                    Precio:
-                    {edit.s === "price" ? (
-                      <span>
-                        <input
-                          type="text"
-                          name="price"
-                          onChange={InputHandler}
-                        />
-                        <button onClick={SendCange}>OK</button>
-                      </span>
-                    ) : (
-                      details.price
-                    )}
-                  </td>
-                  <td>
-                    {edit.e ? (
-                      <button name="price" onClick={EditionPDetail}>
-                        {" "}
-                        🖊{" "}
-                      </button>
-                    ) : (
-                      ""
-                    )}
-                  </td>
-                </tr>
-                <tr>
-                  <td>
-                    Stock:
-                    {edit.s === "stock" ? (
-                      <span>
-                        <input
-                          type="text"
-                          name="stock"
-                          onChange={InputHandler}
-                        />
-                        <button onClick={SendCange}>OK</button>
-                      </span>
-                    ) : (
-                      details.stock
-                    )}
-                  </td>
-                  <td>
-                    {edit.e ? (
-                      <button name="stock" onClick={EditionPDetail}>
-                        {" "}
-                        🖊{" "}
-                      </button>
-                    ) : (
-                      ""
-                    )}
-                  </td>
-                </tr>
-                <tr>Vendedor: {details.userid}</tr>
-                <tr>Disponible: NYI</tr>
-              </td>
-            </tr>
-          </table>
+          <Detail 
+          userId={userId}
+          handleEdition={handleEdition}
+          handleDelete={handleDelete}
+          details={details}
+          currentImg={currentImg}
+          setCurrentImg={setCurrentImg}
+          InputHandler={InputHandler}
+          SendCange={SendCange}
+          EditionPDetail={EditionPDetail}
+          edit={edit}
+          />
           <table className={style.detailTable}>
             <tr>
               <th colSpan="2">Preguntas:</th>
